Move all crypto related code to agg_kzg module
